MyApps-13: Move query implementation, api and dependency injection to submodules.
