+++ app/src/main/java/xyz/torquato/myapps/domain/web/content/GetBooksUseCase.kt
@@ -8,11 +8,11 @@ import kotlinx.coroutines.flow.flatMapMerge
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.scan
-import xyz.torquato.myapps.api.web.IQueryRepository
 import xyz.torquato.myapps.api.web.cache.CacheItem
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.domain.web.content.model.QueryResultCollection
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
+import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
 import xyz.torquato.myaps.domain.api.web.model.BookItem
 import xyz.torquato.myaps.domain.api.web.model.QueryRequest
