+++ data/di/src/main/java/xyz/torquato/myapps/data/di/RepositoryImpl.kt
@@ -7,10 +7,12 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.data.impl.cypher.CypherRepository
 import xyz.torquato.myapps.data.impl.paging.PagingRepository
 import xyz.torquato.myapps.data.impl.waves.SoundRepository
+import xyz.torquato.myapps.data.impl.web.QueryRepository
 import xyz.torquato.myapps.data.impl.web.WebRepository
 import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.sound.ISoundRepository
+import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
 import javax.inject.Singleton
 
@@ -32,5 +34,9 @@ abstract class RepositoryImpl {
     @Binds
     @Singleton
     abstract fun bindPagingImpl(impl: PagingRepository): IPagingRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindQueryImpl(impl: QueryRepository): IQueryRepository
 
 }
